passando para float

+++ pedidos/api/viewsets/pedido_viewset.py
@@ -72,9 +72,9 @@ class PedidosViewSet(viewsets.ModelViewSet):
         # Calcula o valor total do pedido com a taxa de atendimento
         subtotal = 0.0
         for item_pedido in pedido.itenspedido_set.all():
-            subtotal += item_pedido.quantidade * item_pedido.preco_item_mais_complementos
+            subtotal += float(item_pedido.quantidade * item_pedido.preco_item_mais_complementos)
 
-        taxa_atendimento = subtotal * (pedido.restaurante.taxa_serviço / 100.0)
+        taxa_atendimento = float(subtotal) * (float(pedido.restaurante.taxa_serviço) / 100.0)
             
 
         # Cria uma lista de pedido criando chave no stripe
